+++ nextgisweb/core/command.py
@@ -291,13 +291,31 @@ class MaintenanceCommand(Command):
 
     @classmethod
     def argparser_setup(cls, parser, env):
-        pass
+        parser.add_argument(
+            '--estimate-storage', dest='estimate_storage', action='store_true',
+            help='Run storage.estimate after maintenance')
 
     @classmethod
     def execute(cls, args, env):
         for comp in env.chain('maintenance'):
             logger.debug("Maintenance for component: %s...", comp.identity)
             comp.maintenance()
+        
+        if args.estimate_storage:
+            env.core.estimate_storage_all()
+
+
+@Command.registry.register
+class StorageEstimateCommand(Command):
+    identity = 'storage.estimate'
+
+    @classmethod
+    def argparser_setup(cls, parser, env):
+        pass
+
+    @classmethod
+    def execute(cls, args, env):
+        env.core.estimate_storage_all()
 
 
 @Command.registry.register