+++ nextgisweb/core/migration/2e803d47-add-storage-tables.rw.sql
@@ -1,0 +1,8 @@
+/*** { "revision": "2e803d47" } ***/
+
+DROP TABLE core_storage_stat_delta;
+DROP TABLE core_storage_stat_delta_total;
+DROP TABLE core_storage_stat_dimension;
+DROP TABLE core_storage_stat_dimension_total;
+
+DROP FUNCTION core_storage_stat_delta_after_insert();