+++ nextgisweb/feature_attachment/__init__.py
@@ -2,9 +2,17 @@
 from __future__ import division, absolute_import, print_function, unicode_literals
 
 from ..component import Component, require
-from .model import Base
+from ..core import KindOfData
+
+from .model import Base, FeatureAttachment
+from .util import _
 
 __all__ = ['FeatureAttachmentComponent', ]
+
+
+class FeatureAttachmentData(KindOfData):
+    identity = 'feature_attachment'
+    display_name = _("Feature attachments")
 
 
 class FeatureAttachmentComponent(Component):
@@ -18,3 +26,7 @@ class FeatureAttachmentComponent(Component):
     def setup_pyramid(self, config):
         from . import api
         api.setup_pyramid(self, config)
+
+    def estimate_storage(self):
+        for obj in FeatureAttachment.query():
+            yield FeatureAttachmentData, obj.resource_id, obj.size