+++ nextgisweb/pyramid/view.py
@@ -115,6 +115,13 @@ def pkginfo(request):
         'pyramid.control_panel.sysinfo'))
 
 
+def storage(request):
+    request.require_administrator()
+    return dict(
+        title=_("Storage"),
+        dynmenu=request.env.pyramid.control_panel)
+
+
 def backup_browse(request):
     if not request.env.pyramid.options['backup.download']:
         raise HTTPNotFound()
@@ -382,7 +389,7 @@ def setup_pyramid(comp, config):
         .add_view(control_panel, renderer=ctpl('control_panel'))
 
     config.add_route('pyramid.favicon', '/favicon.ico').add_view(favicon)
-    
+
     config.add_route(
         'pyramid.control_panel.sysinfo',
         '/control-panel/sysinfo'
@@ -392,6 +399,12 @@ def setup_pyramid(comp, config):
         'pyramid.control_panel.pkginfo',
         '/control-panel/pkginfo'
     ).add_view(pkginfo)
+
+    if env.core.options['storage.enabled']:
+        config.add_route(
+            'pyramid.control_panel.storage',
+            '/control-panel/storage'
+        ).add_view(storage, renderer=ctpl('storage'))
 
     config.add_route(
         'pyramid.control_panel.backup.browse',
@@ -461,6 +474,10 @@ def setup_pyramid(comp, config):
         dm.Link('settings/home_path', _("Home path"), lambda args: (
             args.request.route_url('pyramid.control_panel.home_path'))),
     )
+
+    if env.core.options['storage.enabled']:
+        comp.control_panel.add(dm.Link('info/storage', _("Storage"), lambda args: (
+            args.request.route_url('pyramid.control_panel.storage'))))
 
     if comp.options['backup.download']:
         comp.control_panel.add(dm.Link(