+++ nextgisweb/raster_layer/__init__.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
 from __future__ import division, absolute_import, print_function, unicode_literals
-import os
-import os.path
+
 from ..component import Component
 
-from .model import Base, RasterLayer
-from .gdaldriver import GDAL_DRIVER_NAME_2_EXPORT_FORMATS
 from . import command  # NOQA
+from .gdaldriver import GDAL_DRIVER_NAME_2_EXPORT_FORMATS
+from .kind_of_data import RasterLayerData
+from .model import Base, RasterLayer, estimate_raster_layer_data
 
 __all__ = ['RasterLayerComponent', 'RasterLayer']
 
@@ -40,3 +40,8 @@ class RasterLayerComponent(Component):
             resource.build_overview(missing_only=True)
 
         # TODO: Cleanup raster_layer directory same as file_storage
+
+    def estimate_storage(self):
+        for resource in RasterLayer.query():
+            size = estimate_raster_layer_data(resource)
+            yield RasterLayerData, resource.id, size