+++ nextgisweb/vector_layer/model.py
@@ -64,7 +64,8 @@ from ..feature_layer import (
     on_data_change,
     query_feature_or_not_found)
 
-from .util import _
+from .kind_of_data import VectorLayerData
+from .util import _, COMP_ID
 
 
 GEOM_TYPE_DB = (
@@ -89,6 +90,15 @@ FIELD_TYPE_DB = (
     db.Date,
     db.Time,
     db.DateTime)
+
+FIELD_TYPE_SIZE = {
+    FIELD_TYPE.INTEGER: 4,
+    FIELD_TYPE.BIGINT: 8,
+    FIELD_TYPE.REAL: 8,
+    FIELD_TYPE.DATE: 4,
+    FIELD_TYPE.TIME: 8,
+    FIELD_TYPE.DATETIME: 8,
+}
 
 FIELD_FORBIDDEN_NAME = ("id", "geom")
 
@@ -447,6 +457,18 @@ class TableInfo(object):
 
         errors = []
 
+        def utf8len(s):
+            return len(s.encode('utf-8'))
+
+        static_size = FIELD_TYPE_SIZE[FIELD_TYPE.INTEGER]
+        dynamic_size = 0
+        string_fields = []
+        for f in self.fields:
+            if f.datatype == FIELD_TYPE.STRING:
+                string_fields.append(f.keyname)
+            else:
+                static_size += FIELD_TYPE_SIZE[f.datatype]
+
         max_fid = None
         for i, feature in enumerate(ogrlayer, start=1):
             if len(errors) >= error_limit:
@@ -468,8 +490,8 @@ class TableInfo(object):
             if geom.GetGeometryType() in (ogr.wkbGeometryCollection, ogr.wkbGeometryCollection25D) \
                and fix_errors != ERROR_FIX.NONE:
                 geom_candidate = None
-                for i in range(geom.GetGeometryCount()):
-                    col_geom = geom.GetGeometryRef(i)
+                for j in range(geom.GetGeometryCount()):
+                    col_geom = geom.GetGeometryRef(j)
                     col_gtype = col_geom.GetGeometryType()
                     if col_gtype not in GEOM_TYPE_OGR:
                         continue
@@ -594,42 +616,42 @@ class TableInfo(object):
                 #     continue
 
             fld_values = dict()
-            for i in range(feature.GetFieldCount()):
-                if i == self.fid_field_index:
+            for k in range(feature.GetFieldCount()):
+                if k == self.fid_field_index:
                     continue
-                fld_type = feature.GetFieldDefnRef(i).GetType()
+                fld_type = feature.GetFieldDefnRef(k).GetType()
 
-                if (not feature.IsFieldSet(i) or feature.IsFieldNull(i)):
+                if (not feature.IsFieldSet(k) or feature.IsFieldNull(k)):
                     fld_value = None
                 elif fld_type == ogr.OFTInteger:
-                    fld_value = feature.GetFieldAsInteger(i)
+                    fld_value = feature.GetFieldAsInteger(k)
                 elif fld_type == ogr.OFTInteger64:
-                    fld_value = feature.GetFieldAsInteger64(i)
+                    fld_value = feature.GetFieldAsInteger64(k)
                 elif fld_type == ogr.OFTReal:
-                    fld_value = feature.GetFieldAsDouble(i)
+                    fld_value = feature.GetFieldAsDouble(k)
                 elif fld_type == ogr.OFTDate:
-                    year, month, day = feature.GetFieldAsDateTime(i)[0:3]
+                    year, month, day = feature.GetFieldAsDateTime(k)[0:3]
                     fld_value = date(year, month, day)
                 elif fld_type == ogr.OFTTime:
-                    hour, minute, second = feature.GetFieldAsDateTime(i)[3:6]
+                    hour, minute, second = feature.GetFieldAsDateTime(k)[3:6]
                     fld_value = time(hour, minute, int(second))
                 elif fld_type == ogr.OFTDateTime:
                     year, month, day, hour, minute, second, tz = \
-                        feature.GetFieldAsDateTime(i)
+                        feature.GetFieldAsDateTime(k)
                     fld_value = datetime(year, month, day,
                                          hour, minute, int(second))
                 elif fld_type == ogr.OFTIntegerList:
-                    fld_value = json.dumps(feature.GetFieldAsIntegerList(i))
+                    fld_value = json.dumps(feature.GetFieldAsIntegerList(k))
                 elif fld_type == ogr.OFTInteger64List:
-                    fld_value = json.dumps(feature.GetFieldAsInteger64List(i))
+                    fld_value = json.dumps(feature.GetFieldAsInteger64List(k))
                 elif fld_type == ogr.OFTRealList:
-                    fld_value = json.dumps(feature.GetFieldAsDoubleList(i))
+                    fld_value = json.dumps(feature.GetFieldAsDoubleList(k))
                 elif fld_type == ogr.OFTStringList:
                     # TODO: encoding
-                    fld_value = json.dumps(feature.GetFieldAsStringList(i))
+                    fld_value = json.dumps(feature.GetFieldAsStringList(k))
                 elif fld_type == ogr.OFTString:
                     try:
-                        fld_value = strdecode(feature.GetFieldAsString(i))
+                        fld_value = strdecode(feature.GetFieldAsString(k))
                     except UnicodeDecodeError:
                         errors.append(_(
                             "It seems like declared and actual attributes "
@@ -639,14 +661,19 @@ class TableInfo(object):
                             feat=fid, attr=i))
                         continue
 
-                fld_name = strdecode(feature.GetFieldDefnRef(i).GetNameRef())
+                fld_name = strdecode(feature.GetFieldDefnRef(k).GetNameRef())
                 fld_values[self[fld_name].key] = fld_value
+
+                if fld_name in string_fields and fld_value is not None:
+                    dynamic_size += utf8len(fld_value)
 
             if len(errors) > 0 and not skip_errors:
                 continue
 
+            geom_bytes = bytearray(geom.ExportToWkb(ogr.wkbNDR))
+            dynamic_size += len(geom_bytes)
             obj = self.model(id=fid, geom=ga.elements.WKBElement(
-                bytearray(geom.ExportToWkb(ogr.wkbNDR)), srid=self.srs_id), **fld_values)
+                geom_bytes, srid=self.srs_id), **fld_values)
 
             DBSession.add(obj)
 
@@ -654,11 +681,16 @@ class TableInfo(object):
             detail = '<br>'.join(html_escape(translate(error)) for error in errors)
             raise VE(message=_("Vector layer cannot be written due to errors."), detail=detail)
 
+        num_features = i
+        size = static_size * num_features + dynamic_size
+
         # Set sequence next value
         if max_fid is not None:
             connection = DBSession.connection()
             connection.execute('ALTER SEQUENCE "%s"."%s" RESTART WITH %d' %
                                (self.sequence.schema, self.sequence.name, max_fid + 1))
+
+        return size
 
 
 class VectorLayerField(Base, LayerField):
@@ -671,7 +703,9 @@ class VectorLayerField(Base, LayerField):
     fld_uuid = db.Column(db.Unicode(32), nullable=False)
 
 
-@implementer(IFeatureLayer, IFieldEditableFeatureLayer, IWritableFeatureLayer, IBboxLayer)
+@implementer(
+    IFeatureLayer, IFieldEditableFeatureLayer, IWritableFeatureLayer,
+    IBboxLayer)
 class VectorLayer(Base, Resource, SpatialLayerMixin, LayerFieldsMixin):
     identity = 'vector_layer'
     cls_display_name = _("Vector layer")
@@ -731,8 +765,10 @@ class VectorLayer(Base, Resource, SpatialLayerMixin, LayerFieldsMixin):
     def load_from_ogr(self, ogrlayer, strdecode=lambda x: x,
                       skip_other_geometry_types=False,
                       fix_errors=ERROR_FIX.default, skip_errors=skip_errors_default):
-        self.tableinfo.load_from_ogr(
+        size = self.tableinfo.load_from_ogr(
             ogrlayer, strdecode, skip_other_geometry_types, fix_errors, skip_errors)
+
+        env.core.reserve_storage(COMP_ID, VectorLayerData, value_data_volume=size, resource=self)
 
     def get_info(self):
         return super(VectorLayer, self).get_info() + (
@@ -919,6 +955,35 @@ class VectorLayer(Base, Resource, SpatialLayerMixin, LayerFieldsMixin):
         )
 
         return extent
+
+
+def estimate_vector_layer_data(resource):
+    tableinfo = TableInfo.from_layer(resource)
+    tableinfo.setup_metadata(resource._tablename)
+    table = tableinfo.table
+
+    static_size = FIELD_TYPE_SIZE[FIELD_TYPE.INTEGER]  # ID field size
+    string_columns = []
+    for f in tableinfo.fields:
+        if f.datatype == FIELD_TYPE.STRING:
+            string_columns.append(f.key)
+        else:
+            static_size += FIELD_TYPE_SIZE[f.datatype]
+
+    size_columns = [func.length(func.ST_AsBinary(table.columns.geom)), ]
+    for key in string_columns:
+        size_columns.append(func.coalesce(func.octet_length(table.columns[key]), 0))
+
+    columns = [func.count(1), ] + [func.coalesce(func.sum(c), 0) for c in size_columns]
+
+    query = sql.select(columns)
+    row = DBSession.connection().execute(query).fetchone()
+
+    num_features = row[0]
+    dynamic_size = sum(row[1:])
+    size = static_size * num_features + dynamic_size
+
+    return size
 
 
 # Create vector_layer schema on table creation